Fix styling issues on channel members tab
This fixes some styling things on the channel members tab, and also
does a little refactoring to remove some unnecessary and duplicated
code.

closes #1246
pr #1297
