+++ static/js/components/admin/EditChannelMembersForm.js
@@ -10,33 +10,29 @@ type Props = {
   memberTypeDescription: string
 } & FormProps<AddMemberForm>
 
-export default class EditChannelMembersForm extends React.Component<Props> {
-  render() {
-    const {
-      processing,
-      validation,
-      form,
-      onUpdate,
-      onSubmit,
-      memberTypeDescription
-    } = this.props
+const EditChannelMembersForm = ({
+  processing,
+  validation,
+  form,
+  onUpdate,
+  onSubmit,
+  memberTypeDescription
+}: Props) => (
+  <div className="members">
+    <form className="add-member-form" onSubmit={onSubmit}>
+      <input
+        type="text"
+        onChange={onUpdate}
+        name="email"
+        placeholder={`The email of the ${memberTypeDescription} you want to add`}
+        value={form.email}
+      />{" "}
+      <button type="submit" className="outlined" disabled={processing}>
+        Submit
+      </button>
+    </form>
+    {validationMessage(validation.email)}
+  </div>
+)
 
-    return (
-      <div className="members">
-        <form className="add-member-form" onSubmit={onSubmit}>
-          <input
-            type="textbox"
-            onChange={onUpdate}
-            name="email"
-            placeholder={`The email of the ${memberTypeDescription} you want to add`}
-            value={form.email}
-          />{" "}
-          <button type="submit" disabled={processing}>
-            Submit
-          </button>
-        </form>
-        {validationMessage(validation.email)}
-      </div>
-    )
-  }
-}
+export default EditChannelMembersForm
