+++ static/js/containers/admin/EditChannelContributorsPage.js
@@ -1,20 +1,14 @@
 // @flow
-import React from "react"
 import R from "ramda"
-import { MetaTags } from "react-meta-tags"
 import { connect } from "react-redux"
 
-import Card from "../../components/Card"
 import EditChannelMembersForm from "../../components/admin/EditChannelMembersForm"
-import EditChannelNavbar from "../../components/admin/EditChannelNavbar"
-import MembersNavbar from "../../components/admin/MembersNavbar"
-import MembersList from "../../components/admin/MembersList"
 import withForm from "../../hoc/withForm"
 import withSingleColumn from "../../hoc/withSingleColumn"
+import editChannelMembershipPage from "./EditChannelMembershipPage"
 
 import { newMemberForm } from "../../lib/channels"
 import { configureForm } from "../../lib/forms"
-import { formatTitle } from "../../lib/title"
 import { actions } from "../../actions"
 import { mergeAndInjectProps } from "../../lib/redux_props"
 import { getChannelName } from "../../lib/util"
@@ -27,136 +21,27 @@ import {
   DIALOG_REMOVE_MEMBER
 } from "../../actions/ui"
 
-import type { AddMemberForm, Channel, Member } from "../../flow/discussionTypes"
-import type { FormErrors, WithFormProps } from "../../flow/formTypes"
-import { channelURL } from "../../lib/url"
+import type { FormErrors } from "../../flow/formTypes"
 
 export const CONTRIBUTORS_KEY = "channel:edit:contributors"
+
 const { getForm, actionCreators } = configureForm(
   CONTRIBUTORS_KEY,
   newMemberForm
 )
 
-const shouldLoadData = R.complement(R.allPass([R.eqProps("channelName")]))
 const usernameGetter = R.prop("contributor_name")
 
-type Props = {
-  channel: Channel,
-  loadChannel: () => Promise<Channel>,
-  loadMembers: () => Promise<*>,
-  members: Array<Member>,
-  removeMember: (channel: Channel, email: string) => Promise<*>,
-  memberToRemove: ?Member,
-  dialogOpen: boolean,
-  setDialogVisibility: (visibility: boolean) => void,
-  setDialogData: (data: any) => void,
-  history: Object,
-  setSnackbarMessage: (obj: Object) => void
-} & WithFormProps<AddMemberForm>
-
-export class EditChannelContributorsPage extends React.Component<Props> {
-  componentDidMount() {
-    this.loadData()
-  }
-
-  componentDidUpdate(prevProps: Props) {
-    if (shouldLoadData(prevProps, this.props)) {
-      this.loadData()
-    }
-  }
-
-  loadData = async () => {
-    const { channel, loadMembers, members } = this.props
-
-    if (!members) {
-      loadMembers()
-    }
-    if (!channel) {
-      this.validateModerator()
-    }
-  }
-
-  validateModerator = async () => {
-    const { loadChannel, history } = this.props
-
-    const channel = await loadChannel()
-    if (!channel.user_is_moderator) {
-      history.push(channelURL(channel.name))
-    }
-  }
-
-  removeMember = async (channel: Channel, member: Member) => {
-    const { removeMember, setSnackbarMessage } = this.props
-    await removeMember(channel, usernameGetter(member))
-    setSnackbarMessage({
-      message: `Successfully removed ${String(member.email)} as a contributor`
-    })
-  }
-
-  render() {
-    const {
-      renderForm,
-      form,
-      channel,
-      members,
-      memberToRemove,
-      dialogOpen,
-      setDialogData,
-      setDialogVisibility
-    } = this.props
-
-    if (!channel || !members || !form) {
-      return null
-    }
-
-    const editable = !channel.membership_is_managed
-
-    return (
-      <React.Fragment>
-        <MetaTags>
-          <title>{formatTitle("Edit Channel")}</title>
-        </MetaTags>
-        <EditChannelNavbar channelName={channel.name} />
-        <Card>
-          <MembersNavbar channel={channel} />
-          {!editable ? (
-            <div className="membership-notice">
-              Membership is managed via MicroMasters
-            </div>
-          ) : (
-            renderForm({
-              memberTypeDescription: "contributor"
-            })
-          )}
-          <MembersList
-            channel={channel}
-            removeMember={this.removeMember}
-            editable={editable}
-            members={members}
-            usernameGetter={usernameGetter}
-            memberTypeDescription="contributor"
-            memberToRemove={memberToRemove}
-            dialogOpen={dialogOpen}
-            setDialogData={setDialogData}
-            setDialogVisibility={setDialogVisibility}
-          />
-        </Card>
-      </React.Fragment>
-    )
-  }
-}
-
-const loadMembers = (channelName: string) =>
-  actions.channelContributors.get(channelName)
-const loadChannel = (channelName: string) => actions.channels.get(channelName)
-const addMember = (channel: Channel, email: string) =>
-  actions.channelContributors.post(channel.name, email)
-const removeMember = (channel: Channel, username: string) =>
-  actions.channelContributors.delete(channel.name, username)
 const onSubmitFailure = (): FormErrors<*> => ({
   email: "Error adding new contributor"
 })
-const onSubmit = (channel, { email }) => addMember(channel, email)
+
+export const EditChannelContributorsPage = editChannelMembershipPage(
+  "contributor",
+  "EditChannelContributorsPage",
+  usernameGetter,
+  false
+)
 
 const mapStateToProps = (state, ownProps) => {
   const channelName = getChannelName(ownProps)
@@ -184,13 +69,13 @@ const mapStateToProps = (state, ownProps) => {
 const mergeProps = mergeAndInjectProps(
   (
     { channelName, channel },
-    { loadMembers, loadChannel, onSubmit, formBeginEdit, setSnackbarMessage }
+    { loadMembers, loadChannel, addMember, formBeginEdit, setSnackbarMessage }
   ) => ({
     loadMembers:    () => loadMembers(channelName),
     loadChannel:    () => loadChannel(channelName),
     onSubmitResult: formBeginEdit,
     onSubmit:       async form => {
-      const newMember = await onSubmit(channel, form)
+      const newMember = await addMember(channel.name, form.email)
       setSnackbarMessage({
         message: `Successfully added ${
           newMember.contributor.email
@@ -204,11 +89,10 @@ export default R.compose(
   connect(
     mapStateToProps,
     {
-      loadMembers,
-      loadChannel,
-      addMember,
-      removeMember,
-      onSubmit,
+      loadMembers:   actions.channelContributors.get,
+      loadChannel:   actions.channels.get,
+      addMember:     actions.channelContributors.post,
+      removeMember:  actions.channelContributors.delete,
       setSnackbarMessage,
       setDialogData: (data: any) =>
         setDialogData({ dialogKey: DIALOG_REMOVE_MEMBER, data: data }),
