+++ static/js/containers/admin/EditChannelContributorsPage_test.js
@@ -255,7 +255,7 @@ describe("EditChannelContributorsPage", () => {
 
     const actions = store.getActions()
 
-    assert.deepEqual(actions[actions.length - 1], {
+    assert.deepEqual(actions[actions.length - 2], {
       type:    SET_SNACKBAR_MESSAGE,
       payload: {
         message: `Successfully removed ${String(
@@ -263,7 +263,13 @@ describe("EditChannelContributorsPage", () => {
         )} as a contributor`
       }
     })
+    assert.deepEqual(actions[actions.length - 1], {
+      type:    HIDE_DIALOG,
+      payload: DIALOG_REMOVE_MEMBER
+    })
   })
+
+  //
   ;[true, false].forEach(hasDialog => {
     it(`passes a dialog value of ${String(hasDialog)}`, async () => {
       const { inner } = await render({
