+++ static/scss/button.scss
@@ -17,6 +17,7 @@ a.link-button {
 .mdc-button.cancel-button,
 button.cancel,
 a.link-button.cancel,
+button.outlined,
 a.link-button.outlined {
   background-color: inherit;
   border: 1px solid $rouge;
