+++ static/scss/layout.scss
@@ -52,6 +52,7 @@
 @import "intra-page-nav";
 @import "channel-banner";
 @import "markdown";
+@import "link";
 
 body {
   font-family: "Source Sans Pro", helvetica, arial, sans-serif !important;
